+++ src/convert.js
@@ -42,8 +42,9 @@ convert.xdrToSep7 = function (conf, xdr, options) {
     const passphrase = resolve.networkPassphrase(conf, options.network)
     sep7 += "&network_passphrase=" + encodeURIComponent(passphrase)
   }
-  if (options.callback)
-    sep7 += "&callback=url:" + encode.url(conf, options.callback)
+  if (options.callback) {
+    sep7 += "&callback=" + encode.url(conf, `url:${options.callback}`)
+  }
   // Not part of the standard.
   // if (options.horizon) sep7 +="&horizon=" + encode.url(conf, options.horizon)